Finish basic flow of site

+++ catalog.py
@@ -29,8 +29,9 @@ session = DBSession()
 @app.route('/catalog/')
 def showCategory():
 	categories = session.query(Category).all()
+	categoryItems = session.query(CatalogItem).order_by(CatalogItem.id.desc()).limit(10)
 	# return "This page will show all my categories"
-	return render_template('index.html', categories=categories)
+	return render_template('index.html', categories=categories, categoryName = 'Latest Items', categoryItems=categoryItems)
 
 @app.route('/catalog/<int:catalog_id>')
 @app.route('/catalog/<int:catalog_id>/items')
@@ -43,15 +44,17 @@ def detailCategory(catalog_id):
 	categoryName = category.name
 	categoryId = category.id
 	categoryItems = session.query(CatalogItem).filter_by(category_id = catalog_id).all()
-	test = 1;
-	return render_template('index.html', categories = categories, categoryItems = categoryItems, categoryName = categoryName, categoryId = categoryId, test = test)
+	return render_template('index.html', categories = categories, categoryItems = categoryItems, categoryName = categoryName, categoryId = categoryId)
 
 	
 @app.route('/catalog/<int:catalog_id>/item/<int:item_id>')
 def detailItem(catalog_id, item_id):
 	catalogItem = session.query(CatalogItem).filter_by(id = item_id).first()
 	creator = getUserInfo(catalogItem.user_id)
-	return render_template('detail_item.html', catalog_id=catalog_id, catalogItem = catalogItem, creator=creator)
+	if login_session.get('user_id') == creator.id:
+		canEdit = True
+	else: canEdit = False
+	return render_template('detail_item.html', catalog_id=catalog_id, catalogItem = catalogItem, creator=creator, canEdit=canEdit)
 
 @app.route(
     '/catalog/<int:catalog_id>/item/<int:item_id>/edit',
@@ -83,11 +86,13 @@ def createItem():
 	# Get all categories
 	categories = session.query(Category).all()
 	if request.method == 'POST':
+		print 'show --->' + request.form['category']
 		newItem = CatalogItem(
 		name=request.form['name'],
 			description=request.form['description'],
 			image_url=request.form['image_url'],
-			category_id=request.form['category'])
+			category_id=request.form['category'],
+			user_id=login_session['user_id'])
 		session.add(newItem)
 		session.commit()
 
@@ -100,8 +105,10 @@ def createItem():
            methods=['GET', 'POST'])
 def deleteItem(catalog_id, item_id):
 	if validate_login():
-		return redirect(url_for('showLogin'))	
+		return redirect(url_for('showLogin'))
 	itemToDelete = session.query(CatalogItem).filter_by(id=item_id).one()
+	if itemToDelete.user_id != login_session.get('user_id'):
+		return "<script>function myFunction() {alert('You are not authorized to delete this item. Please create your own item in order to delete.');}</script><body onload='myFunction()''>"
 	if request.method == 'POST':
 		session.delete(itemToDelete)
 		session.commit()
@@ -180,7 +187,7 @@ def gconnect():
     # Store the access token in the session for later use.
     login_session['credentials'] = credentials.access_token
     login_session['gplus_id'] = gplus_id
-    login_session['api_server'] = 'google'
+    login_session['provider'] = 'google'
 
     # Get user info
     userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
@@ -238,41 +245,126 @@ def gdisconnect():
     	response = make_response(json.dumps('Current user not connected.'), 401)
     	response.headers['Content-Type'] = 'application/json'
     	return response
-    print '----->' + access_token
     url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session['credentials']
     h = httplib2.Http()
     result = h.request(url, 'GET')[0]
     print 'result is '
     print result
-    if result['status'] == '200':
-        del login_session['gplus_id']
-    	del login_session['username']
-    	del login_session['email']
-    	del login_session['picture']
-        del login_session['credentials']
-        del login_session['api_server']
-    	response = make_response(json.dumps('Successfully disconnected.'), 200)
-    	response.headers['Content-Type'] = 'application/json'
-    	return response
-    else:
+    if result['status'] != '200':
 		# Error occured, clear session. Otherwise login will be stuck forever
-        login_session.clear()
-    	response = make_response(json.dumps('Failed to revoke token for given user.', 400))
+        #login_session.clear()
+    	response = make_response(json.dumps('Failed to revoke token for given user.', 401))
     	response.headers['Content-Type'] = 'application/json'
     	return response
-		
+
+
+@app.route('/fbconnect', methods=['POST'])
+def fbconnect():
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    access_token = request.data
+    print "access token received %s " % access_token
+
+
+    app_id = json.loads(open('fb_client_secrets.json', 'r').read())[
+        'web']['app_id']
+    app_secret = json.loads(
+        open('fb_client_secrets.json', 'r').read())['web']['app_secret']
+    url = 'https://graph.facebook.com/oauth/access_token?grant_type=fb_exchange_token&client_id=%s&client_secret=%s&fb_exchange_token=%s' % (
+        app_id, app_secret, access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+
+
+    # Use token to get user info from API
+    userinfo_url = "https://graph.facebook.com/v2.8/me"
+    ''' 
+        Due to the formatting for the result from the server token exchange we have to 
+        split the token first on commas and select the first index which gives us the key : value 
+        for the server access token then we split it on colons to pull out the actual token value
+        and replace the remaining quotes with nothing so that it can be used directly in the graph
+        api calls
+    '''
+    token = result.split(',')[0].split(':')[1].replace('"', '')
+
+    url = 'https://graph.facebook.com/v2.8/me?access_token=%s&fields=name,id,email' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    # print "url sent for API access:%s"% url
+    # print "API JSON result: %s" % result
+    data = json.loads(result)
+    login_session['provider'] = 'facebook'
+    login_session['username'] = data["name"]
+    login_session['email'] = data["email"]
+    login_session['facebook_id'] = data["id"]
+
+    # The token must be stored in the login_session in order to properly logout
+    login_session['access_token'] = token
+
+    # Get user picture
+    url = 'https://graph.facebook.com/v2.8/me/picture?access_token=%s&redirect=0&height=200&width=200' % token
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[1]
+    data = json.loads(result)
+
+    login_session['picture'] = data["data"]["url"]
+
+    # see if user exists
+    user_id = getUserID(login_session['email'])
+    if not user_id:
+        user_id = createUser(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += ' " style = "width: 300px; height: 300px;border-radius: 150px;-webkit-border-radius: 150px;-moz-border-radius: 150px;"> '
+
+    flash("Now logged in as %s" % login_session['username'])
+    return output
+
+
+@app.route('/fbdisconnect')
+def fbdisconnect():
+    facebook_id = login_session['facebook_id']
+    # The access token must me included to successfully logout
+    access_token = login_session['access_token']
+    url = 'https://graph.facebook.com/%s/permissions?access_token=%s' % (facebook_id,access_token)
+    h = httplib2.Http()
+    result = h.request(url, 'DELETE')[1]
+    return "you have been logged out"
+
+# Disconnect based on provider
 @app.route('/disconnect')
 def disconnect():
-	print login_session
-	if login_session.get('api_server') == 'google':
-		return gdisconnect()
-	return "error"
+    if 'provider' in login_session:
+        if login_session['provider'] == 'google':
+            gdisconnect()
+            del login_session['gplus_id']
+            del login_session['credentials']
+        if login_session['provider'] == 'facebook':
+            fbdisconnect()
+            del login_session['facebook_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        del login_session['user_id']
+        del login_session['provider']
+        flash("You have successfully been logged out.")
+        return redirect(url_for('showCategory'))
+    else:
+        flash("You were not logged in")
+        return redirect(url_for('showCategory'))
+
 	
 def validate_login():
-	print 'test'
-	print login_session
 	if 'username' not in login_session:
-		print 'test2'
 		return True
 	else: return False
